Unmark HotAvalonia as a development dependency
I really didn't want to do this, but a lot of people expect to be able to simply click "Install" on HotAvalonia in their favorite GUI package manager and have it just work.

Unfortunately, though, it's impossible for me to achieve that while also keeping the semantic correctness of marking HotAvalonia as a dev-only dependency, because apparently, NuGet package managers have more say in how my package should be installed than I do - thanks to the `IncludeAssets` clause, which is used to **exclude** (!) important assets provided by the library.

So, yeah. For people who actually read the documentation, this change won't mean much. But hopefully, those who don't will stop opening issues complaining that they can't install the package.

Closes #39
Closes #40